+++ R/integration.test.r
@@ -68,8 +68,10 @@
 #' @keywords analysis
 #' @author Dean Adams
 #' @return Objects of class "pls" from integration.test return a list of the following:
-#'  \item{r.pls}{The estimate of morphological integration: PLS.corr. The mean of pairwise
-#'  PLS correlations between partitions is used when there are more than two partitions.}
+#'  \item{r.pls}{The estimate of morphological integration: PLS.corr. The weighted mean of pairwise
+#'  PLS correlations between partitions is used when there are more than two partitions. Weights are
+#'  equal based on the shared number of landmarks compared, relative to the total number of landmarks
+#'  (see Collyer and Adams, 2026).}
 #'    \item{r.pls.mat}{The pairwise r.pls, if the number of partitions is greater than 2.}
 #'    \item{P.value}{The empirically calculated P-value from the resampling procedure.}
 #'   \item{Effect.Size}{The multivariate effect size associated with sigma.d.ratio.}
@@ -87,6 +89,9 @@
 #'    \item{A1.matrix}{Left block (matrix) found from A1 (for 2 modules only).}
 #'    \item{A2.matrix}{Right block (matrix) found from A2 (for 2 modules only).}
 #'    \item{permutations}{The number of random permutations used in the resampling procedure.}
+#'    \item{comp_weights}{The weights for pairwise comparisons, if more than two modules are analyzed.  
+#'    The weights are equal to the number of comparisons divided by two, times the ratio of pair module 
+#'    dimensions to the total dimensions of all modules. (Dimensions = p * k.)}
 #'    \item{call}{The match call.}
 #' @references  Bookstein, F. L., P. Gunz, P. Mitteroecker, H. Prossinger, K. Schaefer, and H. Seidler. 
 #'   2003. Cranial integration in Homo: singular warps analysis of the midsagittal plane in ontogeny and 
@@ -98,6 +103,9 @@
 #' @references Adams, D.C. and M.L. Collyer. 2019. Comparing the strength of modular signal, and evaluating 
 #' alternative modular hypotheses, using covariance ratio effect sizes with morphometric data. 
 #' Evolution. 73:2352-2367.
+#' @references Collyer, M. L., & Adams D. C. (2026). Permutational Biometry: Volume
+#' 1, Univariate Data. Iowa State University Digital Press.  NOT YET PUBLISHED.
+#' .
 #' @seealso \code{\link{two.b.pls}}, \code{\link{modularity.test}}, 
 #' \code{\link{phylo.integration}}, and \code{\link{compare.pls}}
 #' @examples
@@ -266,6 +274,12 @@ integration.test <-function(A, A2 = NULL,
     y <- center(x)
     g <- factor(as.numeric(gps))
     gps.combo <- combn(ngps, 2)
+    gp_pk <- as.vector(table(g))
+    sumpk <- sum(gp_pk)
+    ng_comp <- apply(gps.combo, 2, 
+                     function(x) sum(gp_pk[x]))
+    wts <- ngps / 2 * ng_comp / sumpk
+    
     pls.rand <- sapply(1:perms, function(j) {
       step <- j
       if(print.progress) setTxtProgressBar(pb,step)
@@ -284,12 +298,15 @@ integration.test <-function(A, A2 = NULL,
     
     p.vals <- apply(abs(pls.rand), 1, pval)
     Zs <- apply(pls.rand, 1, effect.size)
-    p.val <- pval(colMeans(abs(pls.rand)))
-    Z <- effect.size(colMeans(pls.rand), center=TRUE)
+    
+    wtMeanPLS <- wts * pls.rand
+    p.val <- pval(colMeans(abs(wts * pls.rand)))
+    Z <- effect.size(colMeans(wts * pls.rand), center=TRUE)
     r.pls.mat <- matrix(0, length(nms), length(nms))
     dimnames(r.pls.mat) <- list(nms, nms)
-    r.pls.mat <- as.dist(r.pls.mat)
+    r.pls.mat <- wt.r.pls.mat <- as.dist(r.pls.mat)
     r.pls.mat[1:nrow(pls.rand)] <- pls.rand[, 1]
+    wt.r.pls.mat[1:nrow(pls.rand)] <- wts * pls.rand[, 1]
   }  
   
   step <- perms + 1
@@ -315,13 +332,15 @@ integration.test <-function(A, A2 = NULL,
                 method = "PLS")
   }
   if(ngps > 2){
-    out <- list(r.pls = mean(r.pls.mat), r.pls.mat = r.pls.mat, 
+    out <- list(r.pls = mean(wt.r.pls.mat), r.pls.mat = r.pls.mat, 
                 P.value = p.val, Z = Z,
                 pairwise.P.values = p.vals, pairwise.Z = Zs,
                 random.r = pls.rand, 
                 permutations = iter+1, call=match.call(),
                 method = "PLS")
   }
+  
+  out$comp_weights <- if(ngps > 2) wts else NA
   
   class(out) <- "pls"
   out