+++ NEWS.md
@@ -4,6 +4,10 @@
 
 * Fixed 3D plotting for points in `plotAllSpecimens`.
 
+### NEW FEATURES
+* `pls` objects with pairwise comparisons now use a weighted mean of r_pls
+values to account for disparate module sizes.
+
 # CHANGES IN GEOMORPH VERSION 4.1.0
 
 ### MAJOR CHANGES